+++ Modules/schema.c
@@ -5,7 +5,7 @@
  * $Id$
  */
 
-#include "Python.h"
+#include "common.h"
 
 #include "schema.h"
 #include "ldap_schema.h"
@@ -18,7 +18,7 @@
 */
 PyObject* c_string_array_to_python(char **string_array)
 {
-  int count = 0;
+  Py_ssize_t count = 0;
   char **s;
   PyObject *py_list;
   if (string_array) {
@@ -46,7 +46,7 @@ PyObject* c_string_array_to_python(char **string_array)
  */
 PyObject* schema_extension_to_python(LDAPSchemaExtensionItem **extensions)
 {
-  int count = 0;
+  Py_ssize_t count = 0;
   LDAPSchemaExtensionItem **e;
   PyObject *py_list, *item_tuple;
   if (extensions) {
